refactor: Improve code structure and add doc strings
Fixes: LEARNER-9242
