+++ OpenEdXMobile/res/layout/fragment_my_courses_list.xml
@@ -24,14 +24,17 @@
                     android:layout_height="wrap_content"
                     android:paddingTop="@dimen/widget_margin_half"
                     android:paddingBottom="@dimen/widget_margin_half"
+                    android:visibility="gone"
                     app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
                     tools:itemCount="2"
-                    tools:listitem="@layout/row_course_list" />
+                    tools:listitem="@layout/row_course_list"
+                    tools:visibility="visible" />
 
                 <include
                     android:id="@+id/stateLayout"
                     layout="@layout/layout_error_state"
-                    android:visibility="visible" />
+                    android:visibility="visible"
+                    tools:visibility="gone" />
 
             </FrameLayout>
         </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
